# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NEXUS API
  version: 1.0.0
  description: |
    Ядро магазина цифровых товаров: каталог, заказ, вебхук оплаты, автовыдача ключа.

    Базовый URL локально: `http://127.0.0.1:8180`

    Типичный сценарий: создать заказ → отправить вебхук `paid` (или `POST /api/orders/{id}/pay`) →
    поллить заказ до `delivered` и забрать `delivery_code`.
servers:
  - url: http://127.0.0.1:8180
    description: local
tags:
  - name: health
  - name: catalog
  - name: orders
  - name: payments
  - name: admin
  - name: providers
paths:
  /health:
    get:
      tags: [health]
      summary: Проверка живости
      operationId: health
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusOK"
              example:
                status: ok

  /api/catalog:
    get:
      tags: [catalog]
      summary: Список товаров
      operationId: listCatalog
      parameters:
        - in: query
          name: featured
          schema:
            type: string
            enum: ["1"]
          description: Только 12 SKU из задания, без сгенерированных GEN-*
        - in: query
          name: limit
          schema:
            type: integer
            default: 48
            maximum: 200
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Список
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogResponse"

  /api/catalog/{sku}:
    get:
      tags: [catalog]
      summary: Карточка товара
      operationId: getProduct
      parameters:
        - $ref: "#/components/parameters/SKU"
      responses:
        "200":
          description: Товар
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          $ref: "#/components/responses/Error"

  /api/orders:
    post:
      tags: [orders]
      summary: Создать заказ
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
            examples:
              simple:
                value:
                  sku: STEAM-TOPUP-500
              withId:
                summary: Свой id — если вебхук может прийти раньше заказа
                value:
                  sku: STEAM-TOPUP-500
                  id: ord_00123
      responses:
        "201":
          description: Создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"

  /api/orders/{id}:
    get:
      tags: [orders]
      summary: Заказ по id
      operationId: getOrder
      parameters:
        - $ref: "#/components/parameters/OrderID"
      responses:
        "200":
          description: Заказ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          $ref: "#/components/responses/Error"

  /api/orders/{id}/pay:
    post:
      tags: [payments]
      summary: Эмуляция оплаты
      description: Генерирует event_id и обрабатывает вебхук paid. Для ручной проверки с витрины.
      operationId: simulatePay
      parameters:
        - $ref: "#/components/parameters/OrderID"
      responses:
        "200":
          description: Заказ после приёма оплаты (выдача может ещё идти)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          $ref: "#/components/responses/Error"

  /webhook/payment:
    post:
      tags: [payments]
      summary: Вебхук платежной системы
      description: |
        Контракт из задания. Доставка at-least-once: один и тот же `event_id` обрабатывается один раз.
        Ответ 200 — принято. 5xx — платежка повторит.
        Если заказа ещё нет, событие сохраняется и применится при создании заказа с тем же id.
      operationId: paymentWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentWebhook"
            example:
              event_id: evt_a1b2c3
              order_id: ord_00123
              status: paid
              amount: 500
              currency: RUB
              created_at: "2025-01-01T12:00:00Z"
      responses:
        "200":
          description: Принято
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusAccepted"
        "400":
          $ref: "#/components/responses/Error"

  /admin/reconcile:
    get:
      tags: [admin]
      summary: Сверка
      operationId: reconcile
      responses:
        "200":
          description: Отчёт
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReconcileReport"

  /admin/orders/{id}/retry:
    post:
      tags: [admin]
      summary: Повторная выдача
      operationId: retryDelivery
      parameters:
        - $ref: "#/components/parameters/OrderID"
      responses:
        "200":
          description: Заказ после попытки
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          $ref: "#/components/responses/Error"

  /admin/recovery/tick:
    post:
      tags: [admin]
      summary: Один тик фонового восстановления
      operationId: recoveryTick
      responses:
        "200":
          description: Запущен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusOK"

  /admin/providers/config:
    get:
      tags: [admin]
      summary: Текущие доли ошибок поставщиков
      operationId: getProviderConfig
      responses:
        "200":
          description: Конфиг
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvidersConfig"
    post:
      tags: [admin]
      summary: Задать доли 5xx и таймаутов
      description: |
        `error_rate` — доля явных 5xx без резерва ключа.
        `timeout_rate` — доля зависаний после резерва (таймаут ≠ отказ).
      operationId: setProviderConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProvidersConfig"
            examples:
              clean:
                value:
                  a: { error_rate: 0, timeout_rate: 0 }
                  b: { error_rate: 0, timeout_rate: 0 }
              fallback:
                value:
                  a: { error_rate: 1, timeout_rate: 0 }
                  b: { error_rate: 0, timeout_rate: 0 }
              timeout:
                value:
                  a: { error_rate: 0, timeout_rate: 1 }
                  b: { error_rate: 1, timeout_rate: 1 }
      responses:
        "200":
          description: Применено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvidersConfig"

  /admin/inventory/restock:
    post:
      tags: [admin]
      summary: Пополнить пул ключей
      operationId: restock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestockRequest"
            example:
              sku: STEAM-TOPUP-500
              codes: [AAAA-BBBB-CCCC, DDDD-EEEE-FFFF]
      responses:
        "200":
          description: Ок
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusOK"
        "400":
          $ref: "#/components/responses/Error"

  /internal/providers/a/issue:
    post:
      tags: [providers]
      summary: Заглушка поставщика A
      operationId: issueA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueRequest"
      responses:
        "200":
          description: Код выдан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueOK"
        "409":
          description: Нет ключей
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueError"
        "502":
          description: Сбой заглушки
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueError"

  /internal/providers/b/issue:
    post:
      tags: [providers]
      summary: Заглушка поставщика B
      operationId: issueB
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueRequest"
      responses:
        "200":
          description: Код выдан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueOK"
        "409":
          description: Нет ключей
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueError"
        "502":
          description: Сбой заглушки
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueError"

components:
  parameters:
    SKU:
      in: path
      name: sku
      required: true
      schema:
        type: string
        example: STEAM-TOPUP-500
    OrderID:
      in: path
      name: id
      required: true
      schema:
        type: string
        example: ord_a1b2c3d4e5f6g7h8

  responses:
    Error:
      description: Ошибка
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  schemas:
    StatusOK:
      type: object
      properties:
        status:
          type: string
          example: ok
    StatusAccepted:
      type: object
      properties:
        status:
          type: string
          example: accepted
    Error:
      type: object
      properties:
        error:
          type: string
          enum: [not_found, invalid, conflict, out_of_stock, internal]
        message:
          type: string
    Product:
      type: object
      properties:
        sku:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [topup, key, subscription, giftcard]
        price:
          type: integer
          description: Цена в RUB, целые рубли
        currency:
          type: string
          example: RUB
        image:
          type: string
        stock:
          type: integer
    CatalogResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        count:
          type: integer
    CreateOrderRequest:
      type: object
      required: [sku]
      properties:
        sku:
          type: string
        id:
          type: string
          description: Необязательный id заказа
    Order:
      type: object
      properties:
        id:
          type: string
        sku:
          type: string
        amount:
          type: integer
        currency:
          type: string
        status:
          type: string
          enum:
            - created
            - paid
            - delivering
            - delivered
            - payment_failed
            - out_of_stock
            - delivery_failed
        delivery_code:
          type: string
          nullable: true
        delivery_request_id:
          type: string
          nullable: true
        delivery_provider:
          type: string
          nullable: true
          enum: [A, B]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PaymentWebhook:
      type: object
      required: [event_id, order_id, status]
      properties:
        event_id:
          type: string
        order_id:
          type: string
        status:
          type: string
          enum: [paid, failed]
        amount:
          type: integer
        currency:
          type: string
        created_at:
          type: string
          format: date-time
    ReconcileReport:
      type: object
      properties:
        paid_not_delivered:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        delivered_not_paid:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        stuck_delivering:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        ledger:
          $ref: "#/components/schemas/LedgerBalance"
    LedgerBalance:
      type: object
      properties:
        debit:
          type: integer
        credit:
          type: integer
        balanced:
          type: boolean
    ProviderRates:
      type: object
      properties:
        error_rate:
          type: number
          minimum: 0
          maximum: 1
        timeout_rate:
          type: number
          minimum: 0
          maximum: 1
    ProvidersConfig:
      type: object
      properties:
        a:
          $ref: "#/components/schemas/ProviderRates"
        b:
          $ref: "#/components/schemas/ProviderRates"
    RestockRequest:
      type: object
      required: [sku, codes]
      properties:
        sku:
          type: string
        codes:
          type: array
          items:
            type: string
    IssueRequest:
      type: object
      required: [request_id, sku, order_id]
      properties:
        request_id:
          type: string
          example: req_ord_00123-a
        sku:
          type: string
        order_id:
          type: string
    IssueOK:
      type: object
      properties:
        status:
          type: string
          example: ok
        request_id:
          type: string
        code:
          type: string
    IssueError:
      type: object
      properties:
        status:
          type: string
          example: error
        reason:
          type: string
          example: out_of_stock